Use usize in route_back_cache.rs instead of doing casting to u64

+++ chain/network/src/routing/routing.rs
@@ -25,9 +25,9 @@ use crate::routing::route_back_cache::RouteBackCache;
 use crate::utils::cache_to_hashmap;
 
 const ANNOUNCE_ACCOUNT_CACHE_SIZE: usize = 10_000;
-const ROUTE_BACK_CACHE_SIZE: u64 = 100_000;
+const ROUTE_BACK_CACHE_SIZE: usize = 100_000;
 const ROUTE_BACK_CACHE_EVICT_TIMEOUT: Duration = Duration::from_millis(120_000);
-const ROUTE_BACK_CACHE_REMOVE_BATCH: u64 = 100;
+const ROUTE_BACK_CACHE_REMOVE_BATCH: usize = 100;
 const PING_PONG_CACHE_SIZE: usize = 1_000;
 const ROUND_ROBIN_MAX_NONCE_DIFFERENCE_ALLOWED: usize = 10;
 const ROUND_ROBIN_NONCE_CACHE_SIZE: usize = 10_000;
